+++ lib/commanded/aggregates/default_lifespan.ex
@@ -1,14 +1,16 @@
 defmodule Commanded.Aggregates.DefaultLifespan do
   @moduledoc """
-  The default implementation of the `Commanded.Aggregates.AggregateLifespan` behaviour.
+  The default implementation of the `Commanded.Aggregates.AggregateLifespan`
+  behaviour.
 
-  It will ensure that an aggregate instance process runs indefinitely once started.
+  It will ensure that an aggregate instance process runs indefinitely once
+  started.
   """
 
   @behaviour Commanded.Aggregates.AggregateLifespan
 
   @doc """
-  Aggregate will run indefinitely once started
+  Aggregate will run indefinitely once started.
   """
-  def after_command(_command), do: :infinity
+  def after_event(_event), do: :infinity
 end